cropai_app/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pandas
numpy
scikit-learn
plotly
joblib
anthropic
python-dotenv
